+++ src/main/kotlin/dev/furq/holodisplays/commands/ReloadCommand.kt
@@ -3,6 +3,9 @@ package dev.furq.holodisplays.commands
 import com.mojang.brigadier.builder.LiteralArgumentBuilder
 import dev.furq.holodisplays.config.ConfigManager
 import dev.furq.holodisplays.handlers.HologramHandler
+import dev.furq.holodisplays.handlers.PacketHandler
+import dev.furq.holodisplays.handlers.TickHandler
+import dev.furq.holodisplays.handlers.ViewerHandler
 import dev.furq.holodisplays.utils.CommandUtils.playSuccessSound
 import net.minecraft.server.command.CommandManager
 import net.minecraft.server.command.ServerCommandSource
@@ -13,7 +16,9 @@ object ReloadCommand {
     fun register(): LiteralArgumentBuilder<ServerCommandSource> = CommandManager.literal("reload")
         .requires { it.hasPermissionLevel(2) }
         .executes { context ->
-            HologramHandler.clearAll()
+            PacketHandler.resetEntityTracking()
+            ViewerHandler.resetHologramObservers()
+            TickHandler.init()
             ConfigManager.reload()
             HologramHandler.reinitialize()
             context.source.sendFeedback({